Only compute closing tag offsets once per scrape
Previously closing tags were computed once per select. If a scraper
used many selects this would lead to wasted work. Now closing tag
offsets are computed once per evaluation of a scraper.

A side-effect of this optimization is that the `select` method changes
types. This method has always been somewhat of an implementation
detail so it is now removed from the public API.