(#27) Split scalpel into two packages
This implements the suggestion in (#27) and creates a new package
`scalpel-core` that does not depend on curl. This new package provides
the Text.HTML.Scalpel.Core module.

The existing `scalpel` package is made to depend on `scalpel` core and
re-exports all symbols from Text.HTML.Scalpel.Core in
Text.HTML.Scalpel.