Fix issue when selecting around invalid HTML
A goal of scalpel is to be able to select locally well formed HTML even
if the entire document is malformed. Prior to this fix, it was not
possible to run the following selector with the given HTML:

  text $ "a" // "c"

  <a> <b> <c> text </c> </a> </b>

This was due to the tree structure scalpel uses internally to represent
parsed HTML. The <a> and <b> nodes would be siblings since neither
encompasses the other with <c> being a child of <b>. When performing a
simple DFS <c> would be missed when matching <a> since it is not in a's
sub-forest. The fix is, when matched, to lift valid sub-forests out of
sibling nodes before continuing the search.