Initial support for scraping multiple sub-trees
The approach utilizes a zipper like API to focus on sibling nodes and
execute scrapers on the focused nodes. Regression tests have been added
that shows that this solves the requirements in issues #41, and #45.

The documentation needs to be cleaned up a bit and I'm not quite happy
with the name `visitSerially` and `visitChildrenSerially`.

Also there is a bad interaction with the newly added textSelector that
makes stepNext/Back behave unexpectedly. Whitespace between tags count
as nodes which leads to unintuitive behavior.

Issue #48