Combine `visitSerially` and `visitChildrenSerially`
The `visitSerially` function was only useful when the entire HTML
document was focused while `visitChildrenSerially` was only useful when
a sub-tree had been focused using `chroot`.

A bit is added to the selection context which notes whether the focus
correspond to the entire original document or if it is a sub-tree. Then
based on this bit the new `inSerial` function performs the behavior of
either `visitSerially` or `visitChildrenSerially`.